backend/requirements.txt
channels==4.1.0
channels_redis==4.2.0
Django==5.0.4
djangorestframework==3.15.1
